+++ bin/ConvertHeaders.py
@@ -18,7 +18,7 @@ from Bio import SeqIO
 # Presto imports
 from presto.Defaults import default_delimiter, default_out_args
 from presto.Annotation import parseAnnotation, flattenAnnotation
-from presto.Commandline import CommonHelpFormatter, getCommonArgParser, parseCommonArgs
+from presto.Commandline import CommonHelpFormatter, checkArgs, getCommonArgParser, parseCommonArgs
 from presto.IO import getFileType, readSeqFile, countSeqFile, getOutputHandle, \
                       printLog, printProgress
 
@@ -506,6 +506,7 @@ if __name__ == '__main__':
     """
     # Parse arguments
     parser = getArgParser()
+    checkArgs(parser)
     args = parser.parse_args()
     args_dict = parseCommonArgs(args)