+++ bin/PairSeq.py
@@ -16,7 +16,7 @@ from Bio import SeqIO
 
 # Presto imports
 from presto.Defaults import choices_coord, default_coord, default_out_args
-from presto.Commandline import CommonHelpFormatter, getCommonArgParser, parseCommonArgs
+from presto.Commandline import CommonHelpFormatter, checkArgs, getCommonArgParser, parseCommonArgs
 from presto.Annotation import parseAnnotation, flattenAnnotation, mergeAnnotation, \
                               getCoordKey
 from presto.IO import getFileType, readSeqFile, countSeqFile, getOutputHandle, \
@@ -232,6 +232,7 @@ if __name__ == '__main__':
     """
     # Parse arguments
     parser = getArgParser()
+    checkArgs(parser)
     args = parser.parse_args()
     args_dict = parseCommonArgs(args)